project/requirements.txt
Django==1.5.2
Markdown==2.4.1
MySQL-python==1.2.5
Pillow==2.4.0
South==0.8.4
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
celery==3.1.12
distribute==0.6.24
django-admin-bootstrapped==1.6.5
django-axes==1.3.6
django-celery==3.1.10
django-classy-tags==0.5.1
django-extensions==1.3.8
django-mptt==0.6.1
django-rosetta==0.7.4
django-sekizai==0.7
eventlet==0.15.0
greenlet==0.4.2
kombu==3.0.20
polib==1.0.4
pygraphviz==1.2
pytz==2014.4
requests==2.3.0
six==1.7.3
sorl-thumbnail==11.12.1b
wiki==0.0.23
wsgiref==0.1.2
